requirements.txt
streamlit>=1.36
gspread>=6
google-auth>=2.30
pandas>=2.2.2
numpy==1.26.4
pyarrow>=14,<15
